+++ chameleonultragui/lib/gui/menu/save_edit.dart
@@ -1,14 +1,9 @@
-import 'dart:convert';
-import 'dart:io';
 
 import 'package:chameleonultragui/gui/menu/card_edit.dart';
 import 'package:chameleonultragui/helpers/mifare_classic/general.dart';
-import 'package:file_picker/file_picker.dart';
-import 'package:file_saver/file_saver.dart';
 import 'package:flutter/material.dart';
 import 'package:chameleonultragui/helpers/general.dart';
 import 'package:chameleonultragui/sharedprefsprovider.dart';
-import 'dart:typed_data';
 import 'package:provider/provider.dart';
 import 'package:chameleonultragui/main.dart';
 
@@ -25,52 +20,6 @@ class SaveEditMenu extends StatefulWidget {
 }
 
 class SaveEditMenuState extends State<SaveEditMenu> {
-  Future<void> saveTag(
-      CardSave tag, ChameleonGUIState appState, bool bin) async {
-    var localizations = AppLocalizations.of(context)!;
-    if (bin) {
-      List<int> tagDump = [];
-      for (var block in tag.data) {
-        tagDump.addAll(block);
-      }
-      try {
-        await FileSaver.instance.saveAs(
-            name: tag.name,
-            bytes: Uint8List.fromList(tagDump),
-            ext: 'bin',
-            mimeType: MimeType.other);
-      } on UnimplementedError catch (_) {
-        String? outputFile = await FilePicker.platform.saveFile(
-          dialogTitle: '${localizations.output_file}:',
-          fileName: '${tag.name}.bin',
-        );
-
-        if (outputFile != null) {
-          var file = File(outputFile);
-          await file.writeAsBytes(Uint8List.fromList(tagDump));
-        }
-      }
-    } else {
-      try {
-        await FileSaver.instance.saveAs(
-            name: tag.name,
-            bytes: const Utf8Encoder().convert(tag.toJson()),
-            ext: 'json',
-            mimeType: MimeType.other);
-      } on UnimplementedError catch (_) {
-        String? outputFile = await FilePicker.platform.saveFile(
-          dialogTitle: '${localizations.output_file}:',
-          fileName: '${tag.name}.json',
-        );
-
-        if (outputFile != null) {
-          var file = File(outputFile);
-          await file.writeAsBytes(const Utf8Encoder().convert(tag.toJson()));
-        }
-      }
-    }
-  }
-
   @override
   Widget build(BuildContext context) {
     var localizations = AppLocalizations.of(context)!;
@@ -115,7 +64,7 @@ class SaveEditMenuState extends State<SaveEditMenu> {
                     if (isMifareClassic(widget.tagSave.tag))
                       ElevatedButton(
                         onPressed: () async {
-                          await saveTag(widget.tagSave, appState, true);
+                          await saveTag(widget.tagSave, context, true);
                           if (context.mounted) {
                             Navigator.pop(context);
                           }
@@ -124,7 +73,7 @@ class SaveEditMenuState extends State<SaveEditMenu> {
                       ),
                     ElevatedButton(
                       onPressed: () async {
-                        await saveTag(widget.tagSave, appState, false);
+                        await saveTag(widget.tagSave, context, false);
                         if (context.mounted) {
                           Navigator.pop(context);
                         }
