+++ chameleonultragui/lib/gui/page/saved_cards.dart
@@ -31,52 +31,6 @@ class SavedCardsPage extends StatefulWidget {
 class SavedCardsPageState extends State<SavedCardsPage> {
   MifareClassicType selectedType = MifareClassicType.m1k;
 
-  Future<void> saveTag(
-      CardSave tag, ChameleonGUIState appState, bool bin) async {
-    var localizations = AppLocalizations.of(context)!;
-    if (bin) {
-      List<int> tagDump = [];
-      for (var block in tag.data) {
-        tagDump.addAll(block);
-      }
-      try {
-        await FileSaver.instance.saveAs(
-            name: tag.name,
-            bytes: Uint8List.fromList(tagDump),
-            ext: 'bin',
-            mimeType: MimeType.other);
-      } on UnimplementedError catch (_) {
-        String? outputFile = await FilePicker.platform.saveFile(
-          dialogTitle: '${localizations.output_file}:',
-          fileName: '${tag.name}.bin',
-        );
-
-        if (outputFile != null) {
-          var file = File(outputFile);
-          await file.writeAsBytes(Uint8List.fromList(tagDump));
-        }
-      }
-    } else {
-      try {
-        await FileSaver.instance.saveAs(
-            name: tag.name,
-            bytes: const Utf8Encoder().convert(tag.toJson()),
-            ext: 'json',
-            mimeType: MimeType.other);
-      } on UnimplementedError catch (_) {
-        String? outputFile = await FilePicker.platform.saveFile(
-          dialogTitle: '${localizations.output_file}:',
-          fileName: '${tag.name}.json',
-        );
-
-        if (outputFile != null) {
-          var file = File(outputFile);
-          await file.writeAsBytes(const Utf8Encoder().convert(tag.toJson()));
-        }
-      }
-    }
-  }
-
   @override
   Widget build(BuildContext context) {
     var appState = context.watch<ChameleonGUIState>();
@@ -406,7 +360,7 @@ class SavedCardsPageState extends State<SavedCardsPage> {
                                     actions: [
                                       ElevatedButton(
                                         onPressed: () async {
-                                          await saveTag(tag, appState, true);
+                                          await saveTag(tag, context, true);
                                           if (context.mounted) {
                                             Navigator.pop(context);
                                           }
@@ -416,7 +370,7 @@ class SavedCardsPageState extends State<SavedCardsPage> {
                                       ),
                                       ElevatedButton(
                                         onPressed: () async {
-                                          await saveTag(tag, appState, false);
+                                          await saveTag(tag, context, false);
                                           if (context.mounted) {
                                             Navigator.pop(context);
                                           }
