+++ chameleonultragui/lib/helpers/general.dart
@@ -1,11 +1,16 @@
+import 'dart:convert';
 import 'dart:io';
 import 'dart:typed_data';
-import 'dart:io' show Platform;
 import 'package:chameleonultragui/bridge/chameleon.dart';
 import 'package:chameleonultragui/connector/serial_abstract.dart';
 import 'package:chameleonultragui/sharedprefsprovider.dart';
+import 'package:file_picker/file_picker.dart';
+import 'package:file_saver/file_saver.dart';
 import 'package:flutter/material.dart';
 import 'package:logger/logger.dart';
+
+// Localizations
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 Future<void> asyncSleep(int milliseconds) async {
   await Future.delayed(Duration(milliseconds: milliseconds));
@@ -226,5 +231,50 @@ AnimationSetting getAnimationModeType(int value) {
     return AnimationSetting.minimal;
   } else {
     return AnimationSetting.none;
+  }
+}
+
+Future<void> saveTag(CardSave tag, BuildContext context, bool bin) async {
+  var localizations = AppLocalizations.of(context)!;
+  if (bin) {
+    List<int> tagDump = [];
+    for (var block in tag.data) {
+      tagDump.addAll(block);
+    }
+    try {
+      await FileSaver.instance.saveAs(
+          name: tag.name,
+          bytes: Uint8List.fromList(tagDump),
+          ext: 'bin',
+          mimeType: MimeType.other);
+    } on UnimplementedError catch (_) {
+      String? outputFile = await FilePicker.platform.saveFile(
+        dialogTitle: '${localizations.output_file}:',
+        fileName: '${tag.name}.bin',
+      );
+
+      if (outputFile != null) {
+        var file = File(outputFile);
+        await file.writeAsBytes(Uint8List.fromList(tagDump));
+      }
+    }
+  } else {
+    try {
+      await FileSaver.instance.saveAs(
+          name: tag.name,
+          bytes: const Utf8Encoder().convert(tag.toJson()),
+          ext: 'json',
+          mimeType: MimeType.other);
+    } on UnimplementedError catch (_) {
+      String? outputFile = await FilePicker.platform.saveFile(
+        dialogTitle: '${localizations.output_file}:',
+        fileName: '${tag.name}.json',
+      );
+
+      if (outputFile != null) {
+        var file = File(outputFile);
+        await file.writeAsBytes(const Utf8Encoder().convert(tag.toJson()));
+      }
+    }
   }
 }
